refactor: result 페이지 metadata 생성 부분 csr로 변경
