+++ stories/dialog.stories.tsx
@@ -13,7 +13,7 @@ const meta = {
       handlers: [
         http.delete('https://vote-server.xyz/api/vote/1', () => {
           return HttpResponse.json({
-            message: 'Vote: 1개, Counts: 0개, Contents: 2개 삭제 되었습니다.',
+            message: 'mocked response',
           })
         }),
       ],
